Move page transition support into library
It's generic enough + allows us to apply bugfixes/changes easily.
