+++ project_template/src/containers/App.js
@@ -2,11 +2,11 @@ import { h } from "preact";
 import { Router as PreactRouter } from "preact-router";
 
 import { history } from "src/support/history";
-import { augmentRouter } from "src/support/pageTransitionSupport";
+import { PageTransitionSupport } from "maji";
 import WelcomePage from "src/modules/welcome/containers/WelcomePage";
 import DetailPage from "src/modules/welcome/containers/DetailPage";
 
-const Router = augmentRouter(PreactRouter);
+const Router = PageTransitionSupport.augmentRouter(PreactRouter, history);
 
 const App = () => (
   <Router history={history}>
