+++ src/cordova/iOSNetworkActivity.js
