+++ src/main.js
@@ -1,6 +1,7 @@
 import I18n from "./i18n";
-import registerCordovaSupport from "./cordova-support";
+import registerCordovaSupport from "./cordovaSupport";
+import * as PageTransitionSupport from "./pageTransitionSupport";
 
 registerCordovaSupport();
 
-export { I18n };
+export { I18n, PageTransitionSupport };
